Add branch filtering for `push` events.
This adds branch filtering for the `push` event so that workflow runs are triggered only for specific branches.

Instances where a pull request has a head branch on the same repository (not originating from a fork) trugger both a `push` and `pull_request:synchronize` event when that branch is updated.

This results in duplicate runs of each workflow. Because of the different triggering event, the `concurrency` configuration does not catch this and both workflows run. This is unnecessary and could result in minutes being wasted when run in private repositories.